# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RankTracker Pro API
  description: |
    # RankTracker Pro - SEO Management Platform API

    A comprehensive SEO management platform that helps businesses track keyword rankings, 
    manage SEO projects, conduct website audits, and monitor competitors.

    ## Features
    - **Multi-tier Subscriptions**: Free, Starter, Professional, Agency plans
    - **Keyword Tracking**: Bulk keyword management with ranking history
    - **SEO Audits**: Comprehensive website analysis
    - **Project Management**: Organize and track multiple SEO projects
    - **Usage Limits**: Real-time monitoring based on subscription plans

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    API requests are limited based on your subscription plan:
    - Free: 10 requests/day
    - Starter: 50 requests/day  
    - Professional: 200 requests/day
    - Agency: 1000 requests/day
  version: 1.0.0
  contact:
    name: RankTracker Pro Support
    url: https://ranktackerpro.com
    email: support@ranktackerpro.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development Server
  - url: https://api.ranktackerpro.com/api/v1
    description: Production Server
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get current user profile information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Users
      summary: Update user profile
      description: Update user profile information
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /users/usage:
    get:
      tags:
        - Users
      summary: Get user subscription usage
      description: Get user subscription usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsageStatistic'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
...
  schemas:
...
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          example: John Smith
        phone:
          type: string
          example: +1234567890
        timezone:
          type: string
          example: Europe/London
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            phone:
              type: string
            timezone:
              type: string
            avatarUrl:
              type: string
            emailVerified:
              type: boolean
            lastLoginAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            subscription:
              $ref: '#/components/schemas/UserSubscription'
...
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        yearlyPrice:
          type: number
        currency:
          type: string
        features:
          type: array
          items:
            type: string
        limits:
          type: object
    UserSubscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, cancelled, expired, trial, suspended, pending]
        plan:
          $ref: '#/components/schemas/SubscriptionPlan'
        billingCycle:
          type: string
          enum: [monthly, yearly]
        startedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    UsageStatistic:
      type: object
      properties:
        type:
          type: string
        current:
          type: integer
        limit:
          type: integer
        percentage:
          type: integer
        resetDate:
          type: string
          format: date
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        details:
          type: array
          items:
            type: string
...